refactor(frontend): use kebab-case source filenames
