+++ apps/desktop/src/renderer/src/app.tsx
@@ -2,7 +2,7 @@ import { SocketIoTransport } from '@linkcode/transport';
 import type { WorkbenchSystemBridge } from '@linkcode/ui';
 import { Workbench, WorkbenchProviders } from '@linkcode/workbench';
 import type { ReactElement } from 'react';
-import { AppI18nProvider } from '@/i18n/AppI18nProvider';
+import { AppI18nProvider } from '@/i18n/app-i18n-provider';
 import { systemBridge } from '@/ipc';
 
 /** The desktop renderer connects to the local daemon (apps/daemon) like every other client. */
