+++ apps/webview/src/app.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 import { RouterProvider } from 'react-router';
-import { RootProviders } from '@/providers/RootProviders';
+import { RootProviders } from '@/providers/root-providers';
 import { appRouter } from '@/routes';
 
 export function App(): ReactElement {
