+++ apps/webview/src/main.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from 'react-dom/client';
-import { App } from '@/App';
+import { App } from '@/app';
 import './index.css';
 
 const el = document.getElementById('root');
