+++ packages/ui/src/chat/index.ts
@@ -1,13 +1,13 @@
-export * from './AssistantMessage';
-export * from './ContentBlockView';
-export * from './ConversationView';
-export * from './DiffBlock';
-export * from './ErrorMessage';
-export * from './Markdown';
-export * from './PermissionCard';
-export * from './PlanCard';
-export * from './TerminalBlock';
-export * from './ThoughtBlock';
-export * from './ToolCallItem';
+export * from './assistant-message';
+export * from './content-block-view';
+export * from './conversation-view';
+export * from './diff-block';
+export * from './error-message';
+export * from './markdown';
+export * from './permission-card';
+export * from './plan-card';
+export * from './terminal-block';
+export * from './thought-block';
+export * from './tool-call-item';
 export * from './types';
-export * from './UserMessage';
+export * from './user-message';
