+++ packages/ui/src/chat/permission-card.tsx
