+++ packages/ui/src/chat/terminal-block.tsx
