+++ packages/ui/src/native/index.ts
@@ -1,1 +1,1 @@
-export * from './MobileHome';
+export * from './mobile-home';
