+++ packages/ui/src/shell/session-item.tsx
